Avoid updates / generating draw nodes for valid BufferedContainers

+++ osu.Framework/Graphics/Containers/CompositeDrawable.cs
@@ -521,12 +521,16 @@ namespace osu.Framework.Graphics.Containers
             if (aliveInternalChildren.Count == 0)
                 return true;
 
-            var childMaskingBounds = ComputeChildMaskingBounds(maskingBounds);
+            if (RequiresChildrenUpdate)
+            {
+                var childMaskingBounds = ComputeChildMaskingBounds(maskingBounds);
 
-            // We iterate by index to gain performance
-            // ReSharper disable once ForCanBeConvertedToForeach
-            for (int i = 0; i < aliveInternalChildren.Count; i++)
-                aliveInternalChildren[i].UpdateSubTreeMasking(this, childMaskingBounds);
+
+                // We iterate by index to gain performance
+                // ReSharper disable once ForCanBeConvertedToForeach
+                for (int i = 0; i < aliveInternalChildren.Count; i++)
+                    aliveInternalChildren[i].UpdateSubTreeMasking(this, childMaskingBounds);
+            }
 
             return true;
         }
@@ -730,7 +734,9 @@ namespace osu.Framework.Graphics.Containers
             }
         }
 
-        internal sealed override DrawNode GenerateDrawNodeSubtree(int treeIndex)
+        internal virtual bool AddChildDrawNodes => true;
+
+        internal override DrawNode GenerateDrawNodeSubtree(int treeIndex)
         {
             // No need for a draw node at all if there are no children and we are not glowing.
             if (aliveInternalChildren.Count == 0 && CanBeFlattened)
@@ -743,13 +749,16 @@ namespace osu.Framework.Graphics.Containers
             if (cNode.Children == null)
                 cNode.Children = new List<DrawNode>(aliveInternalChildren.Count);
 
-            List<DrawNode> target = cNode.Children;
+            if (AddChildDrawNodes)
+            {
+                List<DrawNode> target = cNode.Children;
 
-            int j = 0;
-            addFromComposite(treeIndex, ref j, this, target);
+                int j = 0;
+                addFromComposite(treeIndex, ref j, this, target);
 
-            if (j < target.Count)
-                target.RemoveRange(j, target.Count - j);
+                if (j < target.Count)
+                    target.RemoveRange(j, target.Count - j);
+            }
 
             return cNode;
         }
